requirements.txt
alembic==1.12.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
Flask==2.2.5
Flask-Cors==5.0.0
ftfy==6.1.1
greenlet==3.1.1
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.6
joblib==1.3.2
langdetect==1.0.9
langid==1.1.6
Mako==1.2.4
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.21.6
optuna==4.0.0
packaging==24.0
pandas==1.3.5
python-dateutil==2.9.0.post0
python-whois==0.9.5
pytz==2025.2
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.17.0
SQLAlchemy==2.0.40
threadpoolctl==3.1.0
torch==1.13.1
tqdm==4.67.1
typing_extensions==4.7.1
Unidecode==1.4.0
urllib3==2.0.7
wcwidth==0.2.13
Werkzeug==2.2.3
zipp==3.15.0
